refactor(shell): delete .ssh-agent.sock if exists

+++ shell/.bash_profile
@@ -15,7 +15,7 @@ then
    if [ "$RUNNING_AGENT" = "0" ]
    then
       # Launch a new instance of the agent
-      \rm "$HOME"/.ssh-agent.sock
+      [[ -f "$HOME"/.ssh-agent.sock ]] && \rm "$HOME"/.ssh-agent.sock
       ssh-agent -a "$HOME"/.ssh-agent.sock -s &> "$HOME"/.ssh/ssh-agent
    fi
    eval "$(cat "$HOME"/.ssh/ssh-agent)"
